requirements.txt
Flask==3.1.2
Flask-CORS==6.0.1
reportlab==4.4.4
python-dotenv==1.0.1
openai==1.55.3
gunicorn==21.2.0
# Database
Flask-SQLAlchemy==3.1.1
# Payment providers
stripe==7.0.0
paypalrestsdk==1.13.1
requests==2.31.0

